add docs and tests
