--- docs/source/custom_forms/examples/home/tables.py
@@ -1,6 +1,0 @@
-from piccolo.columns.column_types import Varchar
-from piccolo.table import Table
-
-
-class Task(Table):
-    title = Varchar()
