+++ e2e/test_add_row.py
@@ -1,6 +1,6 @@
 from playwright.sync_api import Page
 
-from piccolo_admin.example import Director
+from piccolo_admin.example.tables import Director
 
 from .pages import AddRowPage, LoginPage
