+++ e2e/test_codegen.py
@@ -48,26 +48,19 @@ def test_custom_form(playwright: Playwright, dev_server) -> None:
     # Open new page
     page = context.new_page()
     page.goto("http://localhost:8000/#/login?nextURL=%2F")
-    page.locator('input[name="username"]').click()
     page.locator('input[name="username"]').fill("piccolo")
-    page.locator('input[name="username"]').press("Tab")
     page.locator('input[name="password"]').fill("piccolo123")
     page.locator('input[name="password"]').press("Enter")
-    page.get_by_role("link", name="Business email form").click()
-    page.locator('input[name="email"]').click()
-    page.locator('input[name="email"]').fill("director@director.com")
-    page.locator('input[name="content"]').click()
-    page.locator('input[name="content"]').fill("Hello from Piccolo Admin")
-    page.get_by_role("button", name="Submit").click()
-    page.get_by_role("paragraph").filter(has_text="Back").get_by_role(
-        "link", name="Back"
-    ).click()
     page.get_by_role("link", name="Booking form").click()
-    page.locator('input[name="email"]').click()
     page.locator('input[name="email"]').fill("customer@customer.com")
-    page.locator('input[name="name"]').click()
     page.locator('input[name="name"]').fill("Bob")
-    page.locator('input[name="notes"]').click()
+    page.locator('input[name="tickets"]').fill("2")
+
+    # We can't fill hidden elements - so need to do this instead:
+    page.evaluate(
+        'document.querySelector("input[name=starts_at]").setAttribute("value", "2024-10-01T12:00:00")'  # noqa: E501
+    )
+
     page.locator('input[name="notes"]').fill("Hello from Piccolo Admin")
     page.get_by_role("button", name="Submit").click()
     page.get_by_role("link", name="Back to home page").click()
@@ -85,9 +78,7 @@ def test_bulk_update(playwright: Playwright, dev_server) -> None:
     # Open new page
     page = context.new_page()
     page.goto("http://localhost:8000/#/login?nextURL=%2F")
-    page.locator('input[name="username"]').click()
     page.locator('input[name="username"]').fill("piccolo")
-    page.locator('input[name="username"]').press("Tab")
     page.locator('input[name="password"]').fill("piccolo123")
     page.locator('input[name="password"]').press("Enter")
     page.get_by_role("link", name="director", exact=True).click()
@@ -113,9 +104,7 @@ def test_custom_links(playwright: Playwright, dev_server) -> None:
     # Open new page
     page = context.new_page()
     page.goto("http://localhost:8000/#/login?nextURL=%2F")
-    page.locator('input[name="username"]').click()
     page.locator('input[name="username"]').fill("piccolo")
-    page.locator('input[name="username"]').press("Tab")
     page.locator('input[name="password"]').fill("piccolo123")
     page.locator('input[name="password"]').press("Enter")
     page.get_by_role("link", name="Top Movies").click()
