+++ e2e/test_csv_download.py
@@ -4,7 +4,7 @@ import tempfile
 
 from playwright.sync_api import Page
 
-from piccolo_admin.example import Movie
+from piccolo_admin.example.tables import Movie
 
 from .pages import LoginPage, RowListingPage
