+++ piccolo_admin/example/forms/__init__.py
@@ -1,0 +1,5 @@
+from .csv import FORM as CSV_FORM
+from .email import FORM as EMAIL_FORM
+from .image import FORM as IMAGE_FORM
+
+FORMS = [CSV_FORM, EMAIL_FORM, IMAGE_FORM]
