+++ piccolo_admin/example/tables.py
@@ -1,26 +1,11 @@
-"""
-An example of how to configure and run the admin.
-
-Can be run from the command line using `python -m piccolo_admin.example`,
-or `admin_demo`.
-"""
-
-import asyncio
-import csv
 import datetime
 import decimal
 import enum
-import io
 import logging
 import os
 import random
-import shutil
-import smtplib
 import typing as t
 
-import targ
-from hypercorn.asyncio import serve
-from hypercorn.config import Config
 from piccolo.apps.user.tables import BaseUser
 from piccolo.columns.column_types import (
     JSON,
@@ -48,25 +33,12 @@ from piccolo.columns.readable import Readable
 from piccolo.engine.postgres import PostgresEngine
 from piccolo.engine.sqlite import SQLiteEngine
 from piccolo.table import Table, create_db_tables_sync, drop_db_tables_sync
-from piccolo_api.encryption.providers import XChaCha20Provider
-from piccolo_api.media.local import LocalMediaStorage
-from piccolo_api.media.s3 import S3MediaStorage
-from piccolo_api.mfa.authenticator.provider import AuthenticatorProvider
 from piccolo_api.mfa.authenticator.tables import (
     AuthenticatorSecret as AuthenticatorSecret_,
 )
 from piccolo_api.session_auth.tables import SessionsBase
-from pydantic import BaseModel, field_validator
-from starlette.requests import Request
 
-from piccolo_admin.endpoints import (
-    FileResponse,
-    FormConfig,
-    OrderBy,
-    TableConfig,
-    create_admin,
-)
-from piccolo_admin.example_data import (
+from piccolo_admin.example.data import (
     DIRECTORS,
     MOVIE_WORDS,
     MOVIES,
@@ -76,65 +48,6 @@ from piccolo_admin.example_data import (
 )
 
 logger = logging.getLogger()
-
-try:
-    """
-    If you want to try out S3, create a .env file in this folder, with the
-    following contents (inserting your S3 credentials where appropriate):
-
-        AWS_ACCESS_KEY_ID=abc123
-        AWS_SECRET_ACCESS_KEY=abc123
-        BUCKET_NAME=bucket123
-
-    These values can also be added if required:
-
-        ENDPOINT_URL=s3.cloudprovider.com
-        REGION_NAME=my-region
-
-    The ``Director.photo`` column will then use S3 for storage.
-
-    """
-
-    import dotenv
-
-    dotenv.load_dotenv(os.path.join(os.path.dirname(__file__), ".env"))
-
-    AWS_ACCESS_KEY_ID = os.environ.get("AWS_ACCESS_KEY_ID")
-    AWS_SECRET_ACCESS_KEY = os.environ.get("AWS_SECRET_ACCESS_KEY")
-    BUCKET_NAME = os.environ.get("BUCKET_NAME")
-
-    USE_S3 = all(
-        (
-            AWS_SECRET_ACCESS_KEY,
-            AWS_SECRET_ACCESS_KEY,
-            BUCKET_NAME,
-        )
-    )
-
-    if USE_S3:
-        logger.info("Using S3")
-
-        S3_CONFIG = {
-            "aws_access_key_id": AWS_ACCESS_KEY_ID,
-            "aws_secret_access_key": AWS_SECRET_ACCESS_KEY,
-        }
-
-        ENDPOINT_URL = os.environ.get("ENDPOINT_URL")
-        if ENDPOINT_URL:
-            S3_CONFIG["endpoint_url"] = ENDPOINT_URL
-
-        REGION_NAME = os.environ.get("REGION_NAME")
-        if REGION_NAME:
-            S3_CONFIG["region_name"] = REGION_NAME
-
-except ImportError:
-    USE_S3 = False
-
-
-MEDIA_ROOT = os.path.join(os.path.dirname(__file__), "example_media")
-if not USE_S3 and not os.path.exists(MEDIA_ROOT):
-    os.mkdir(MEDIA_ROOT)
-
 
 USERNAME = "piccolo"
 PASSWORD = "piccolo123"
@@ -377,126 +290,7 @@ class Choices(Table):
 
 
 ###############################################################################
-
-
-class BusinessEmailModel(BaseModel):
-    email: str
-    title: str = "Enquiry"
-    content: str
-
-    @field_validator("email")
-    def validate_email(cls, v):
-        if "@" not in v:
-            raise ValueError("not valid email")
-        return v
-
-
-class BookingModel(BaseModel):
-    email: str
-    name: str
-    notes: str = "N/A"
-
-    @field_validator("email")
-    def validate_email(cls, v):
-        if "@" not in v:
-            raise ValueError("not valid email")
-        return v
-
-
-class DownloadMoviesModel(BaseModel):
-    director_name: str
-
-
-class DownloadScheduleModel(BaseModel):
-    date: datetime.date
-
-
-def business_email_endpoint(request: Request, data: BusinessEmailModel) -> str:
-    sender = "info@example.com"
-    receivers = [data.email]
-
-    message = f"""From: Piccolo Admin <info@example.com>
-    To: Colleague <{data.email}>
-    Subject: {data.title}
-    {data.content}
-    """
-
-    try:
-        smtpObj = smtplib.SMTP("localhost:1025")
-        smtpObj.sendmail(sender, receivers, message)
-        print("Successfully sent email")
-    except (smtplib.SMTPException, ConnectionRefusedError):
-        print("Error: unable to send email")
-
-    return "Email sent"
-
-
-def booking_endpoint(request: Request, data: BookingModel) -> str:
-    sender = "info@example.com"
-    receivers = [data.email]
-
-    message = f"""From: Bookings <info@example.com>
-    To: To Friend <{data.email}>
-    Subject: {data.name} booking
-    {data.notes}
-    """
-
-    try:
-        smtpObj = smtplib.SMTP("localhost:1025")
-        smtpObj.sendmail(sender, receivers, message)
-        print("Successfully sent email")
-    except (smtplib.SMTPException, ConnectionRefusedError):
-        print("Error: unable to send email")
-
-    return "Booking complete"
-
-
-async def download_movies(
-    request: Request, data: DownloadMoviesModel
-) -> FileResponse:
-    """
-    An example custom form function which downloads a CSV file.
-    """
-    movies = await Movie.select(Movie.name, Movie.release_date).where(
-        Movie.director._.name == data.director_name
-    )
-
-    output_file = io.StringIO(None)
-
-    writer = csv.DictWriter(
-        f=output_file,
-        fieldnames=["name", "release_date"],
-    )
-    writer.writeheader()
-    writer.writerows(movies)
-
-    return FileResponse(
-        contents=output_file,
-        file_name="director_movies.csv",
-        media_type="text/csv",
-    )
-
-
-def download_schedule(
-    request: Request, data: DownloadScheduleModel
-) -> FileResponse:
-    """
-    An example custom form function which downloads an image file.
-    """
-    file_name = "movie_listings.jpg"
-    with open(
-        os.path.join(os.path.dirname(__file__), "files", file_name),
-        "rb",
-    ) as f:
-        output_file = io.BytesIO()
-        shutil.copyfileobj(f, output_file)
-
-    return FileResponse(
-        contents=output_file,
-        file_name=file_name,
-        media_type="image/jpeg",
-    )
-
+# Create the schema and populate data
 
 TABLE_CLASSES: t.Tuple[t.Type[Table], ...] = (
     Director,
@@ -514,189 +308,6 @@ TABLE_CLASSES: t.Tuple[t.Type[Table], ...] = (
     ConstraintTarget,
     DateTimeColumns,
     Choices,
-)
-
-
-movie_config = TableConfig(
-    table_class=Movie,
-    visible_columns=[
-        Movie._meta.primary_key,
-        Movie.name,
-        Movie.rating,
-        Movie.duration,
-        Movie.director,
-        Movie.won_oscar,
-        Movie.poster,
-        Movie.tags,
-        Movie.screenshots,
-    ],
-    visible_filters=[
-        Movie.name,
-        Movie.rating,
-        Movie.director,
-        Movie.duration,
-        Movie.genre,
-        Movie.screenshots,
-        Movie.poster,
-        Movie.release_date,
-    ],
-    rich_text_columns=[Movie.description],
-    media_storage=(
-        LocalMediaStorage(
-            column=Movie.poster,
-            media_path=os.path.join(MEDIA_ROOT, "movie_poster"),
-        ),
-        LocalMediaStorage(
-            column=Movie.screenshots,
-            media_path=os.path.join(MEDIA_ROOT, "movie_screenshots"),
-        ),
-    ),
-    menu_group="Movies",
-    order_by=[OrderBy(Movie.rating, ascending=False)],
-)
-
-director_config = TableConfig(
-    table_class=Director,
-    visible_columns=[
-        Director._meta.primary_key,
-        Director.name,
-        Director.gender,
-        Director.photo,
-    ],
-    media_storage=(
-        (
-            S3MediaStorage(
-                column=Director.photo,
-                bucket_name=t.cast(str, BUCKET_NAME),
-                folder_name="director_photo",
-                connection_kwargs=S3_CONFIG,
-            )
-            if USE_S3
-            else LocalMediaStorage(
-                column=Director.photo,
-                media_path=os.path.join(MEDIA_ROOT, "photo"),
-            )
-        ),
-    ),
-    menu_group="Movies",
-)
-
-studio_config = TableConfig(
-    table_class=Studio,
-    menu_group="Movies",
-)
-
-ticket_config = TableConfig(
-    table_class=Ticket,
-    menu_group="Booking",
-    link_column=Ticket.booked_by,
-    visible_columns=[
-        Ticket.booked_by,
-        Ticket.movie,
-        Ticket.start_date,
-        Ticket.start_time,
-    ],
-    time_resolution={Ticket.start_time: 60, Ticket.booked_on: 1},
-)
-
-array_columns_config = TableConfig(
-    table_class=ArrayColumns, menu_group="Testing"
-)
-
-nullable_config = TableConfig(
-    table_class=NullableColumns,
-    menu_group="Testing",
-)
-
-required_columns_config = TableConfig(
-    table_class=RequiredColumns,
-    menu_group="Testing",
-)
-
-sorted_columns_config = TableConfig(
-    table_class=SortedColumns,
-    order_by=[OrderBy(SortedColumns.integer, ascending=True)],
-    menu_group="Testing",
-)
-
-constraints_config = TableConfig(
-    table_class=Constraints,
-    menu_group="Testing",
-)
-
-constraints_target_config = TableConfig(
-    table_class=ConstraintTarget,
-    menu_group="Testing",
-)
-
-date_time_config = TableConfig(
-    table_class=DateTimeColumns, menu_group="Testing"
-)
-
-choices_config = TableConfig(
-    table_class=Choices,
-    menu_group="Testing",
-)
-
-APP = create_admin(
-    [
-        movie_config,
-        director_config,
-        studio_config,
-        ticket_config,
-        array_columns_config,
-        nullable_config,
-        required_columns_config,
-        sorted_columns_config,
-        constraints_config,
-        constraints_target_config,
-        date_time_config,
-        choices_config,
-    ],
-    forms=[
-        FormConfig(
-            name="Business email form",
-            pydantic_model=BusinessEmailModel,
-            endpoint=business_email_endpoint,
-            description="Send an email to a business associate.",
-        ),
-        FormConfig(
-            name="Booking form",
-            pydantic_model=BookingModel,
-            endpoint=booking_endpoint,
-            description="Make a booking for a customer.",
-        ),
-        FormConfig(
-            name="Download director movies",
-            pydantic_model=DownloadMoviesModel,
-            endpoint=download_movies,
-            description=(
-                "Download a list of movies for the director as a CSV file."
-            ),
-        ),
-        FormConfig(
-            name="Download schedule",
-            pydantic_model=DownloadScheduleModel,
-            endpoint=download_schedule,
-            description=("Download the schedule for the day."),
-        ),
-    ],
-    auth_table=User,
-    session_table=Sessions,
-    sidebar_links={
-        "Top Movies": "/#/movie?__order=-box_office",
-        "Google": "https://google.com",
-    },
-    mfa_providers=[
-        AuthenticatorProvider(
-            encryption_provider=XChaCha20Provider(
-                encryption_key=(
-                    b"\x01\xfdN\xe4E?\xaa\xf8<e\xfc\x9f\x0b9\x8b\x00H%~\xe1/\xd7\xdcz\xff\xd88\xdajd\xae\x06"  # noqa: E501
-                )
-            ),
-            secret_table=AuthenticatorSecret,
-        ),
-    ],
 )
 
 
@@ -842,41 +453,3 @@ def populate_data(inflate: int = 0, engine: str = "sqlite"):
                     movies.append(movie)
 
                 Movie.insert(*movies).run_sync()
-
-
-def run(persist: bool = False, engine: str = "sqlite", inflate: int = 0):
-    """
-    Start the Piccolo admin.
-
-    :param persist:
-        If True, we don't rebuild all of the data each time.
-    :param engine:
-        Options are sqlite and postgres. By default sqlite is used.
-    :param inflate:
-        If set, this number of extra rows are inserted containing dummy data.
-        This is useful when you need to test with lots of data. Example
-        `--inflate=10000`.
-
-    """
-    set_engine(engine)
-    create_schema(persist=persist)
-
-    if not persist:
-        populate_data(inflate=inflate, engine=engine)
-
-    # Server
-    class CustomConfig(Config):
-        use_reloader = True
-        accesslog = "-"
-
-    asyncio.run(serve(APP, CustomConfig()))
-
-
-def main():
-    cli = targ.CLI(description="Piccolo Admin")
-    cli.register(run)
-    cli.run(solo=True)
-
-
-if __name__ == "__main__":
-    main()
